requirements.txt
django
gunicorn
whitenoise
mssql-django
cryptography
django-encrypted-model-fields
django-encrypted-files
Pillow

djangorestframework
django-cors-headers
django-filter
djangorestframework-simplejwt

python-dotenv
requests
